Partial completed
In the example we implemented security team which will monitor the external visitors of the company, and related employees that connect to those visitors. these two will be the observer to the Hub which monitors all appearence of external visitors. we implemented abstract class to follow the DRY principle, using interfaces to give highly readable code, using padding for format. and using observer pattern for subscribe and unsubscribe methods